backend/requirements.txt
# FastAPI and ASGI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Redis
redis>=5.0.1
celery>=5.3.4

# AWS S3
boto3>=1.34.0

# PDF Generation
reportlab>=4.0.7

# Razorpay
razorpay>=1.3.0

# Security & Rate Limiting
slowapi>=0.1.9
prometheus-fastapi-instrumentator>=5.10.0

# HTTP Client
httpx>=0.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
aiosqlite>=0.19.0

# Sentry (Optional - for error tracking)
sentry-sdk>=2.0.0

# Development
python-dotenv>=1.0.0
black>=23.12.0
isort>=5.13.0
mypy>=1.8.0
openpyxl>=3.1.2
prometheus-client>=0.19.0

